requirements-test.txt
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.35
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
openai==1.51.0
fastapi==0.115.0
aiokafka==0.11.0
